requirements.txt
pandas
numpy
ta
python-binance
websocket-client
pytz
python-dotenv
requests
